Fix race-condition in download
Previously this was waiting for the `end` event on the read stream.
Instead it should wait for the `close` event on the file stream. This is
important because after we get to the end of the read stream, we're
still writing the changes to disk on the file/write stream.